Refactor : 모든 controller 테스트에 parent 적용
- 모든 컨트롤러 테스트에 ControllerTestParent 를 상속받도록 하여서 중복을 제거하였습니다.
